pyproject.toml
[tool.poetry]
name = "adboard"
version = "0.1.0"
description = ""
authors = ["Alexandra Goldobina <goldobina.s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
black = "^24.10.0"
isort = "^5.13.2"
ipython = "^8.30.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
